Project/requirements.txt
google-cloud-storage
google-cloud-speech
janome
